Refactor the routing code into multiple files
This should make it easier to manage and allow a new routing matcher
to be developed.
